Create MevBot.sol
The BOT sniffs the DEX Mempool for transactions with high slippage, determining if a sandwich attack would be profitable. Bots then compete to buy up the token on-chain as swiftly as possible, sandwiching the victim’s transaction and creating a profitable slippage opportunity. The bot always adds 1 gas more than everybody else’s, as long as it remains profitable, ensuring a large number of profitable transactions. It then sends back the ETH/BNB to the contract ready for withdrawal. This bot performs all these tasks faster than 99% of other bots out there